CI: Show diff in bytes instead of percentage
